Simplify cfg usage and make it less noisy
