+++ core/Cargo.toml
@@ -135,6 +135,7 @@ aegis = "0.9.5"
 [build-dependencies]
 chrono = { workspace = true, default-features = false }
 built = { version = "0.7.5", features = ["chrono"] }
+cfg_aliases = "0.2.1"
 
 [target.'cfg(not(target_family = "windows"))'.dev-dependencies]
 pprof = { version = "0.14.0", features = ["criterion", "flamegraph"] }
